+++ Content.Server/Ganimed/Disease/Effects/EffectSystem.cs
@@ -26,7 +26,7 @@ public sealed partial class DiseaseEffectSystem : SharedDiseaseEffectSystem
         SubscribeLocalEvent<DiseaseCarrierComponent, DiseaseEffectArgs<DiseasePopUp>>(DiseasePopUp);
         SubscribeLocalEvent<DiseaseCarrierComponent, DiseaseEffectArgs<DiseaseSnough>>(DiseaseSnough);
         SubscribeLocalEvent<DiseaseCarrierComponent, DiseaseEffectArgs<DiseaseVomit>>(DiseaseVomit);
-        SubscribeLocalEvent<DiseaseCarrierComponent, DiseaseEffectArgs<DiseaseCyborgConversion>>(DiseaseCyborgConversion);
+        SubscribeLocalEvent<DiseaseCarrierComponent, DiseaseEffectArgs<DiseaseCatConversion>>(DiseaseCatConversion);
 
         SubscribeLocalEvent<DiseaseInfectionSpreadEvent>(OnSpreadEvent);
     }
