+++ Resources/Locale/ru-RU/ganimed/disease/diagnoser.ftl
@@ -1,0 +1,20 @@
+﻿diagnoser-cant-use-swab = {CAPITALIZE(THE($machine))} отвергает {THE($swab)}.
+diagnoser-insert-swab = Вы помещаете {THE($swab)} в {THE($machine)}.
+diagnoser-disease-report = Отчёт о заболевании: {CAPITALIZE($disease)}
+diagnoser-disease-report-none = Чек о хорошем состоянии
+diagnoser-disease-report-none-contents = [color=green]В этом месте небыло обнаружено болезней.[/color]
+diagnoser-disease-report-name = Имя болезни: {CAPITALIZE($disease)}
+diagnoser-disease-report-infectious = Заражение: [color=red]Да[/color]
+diagnoser-disease-report-not-infectious = Заражение: [color=green]Нет[/color]
+diagnoser-disease-report-cureresist-none = Сопротивление космоцилину: [color=green]Отсутствует[/color]
+diagnoser-disease-report-cureresist-low = Сопротивление космоцилину: [color=yellow]Слабое[/color]
+diagnoser-disease-report-cureresist-medium = Сопротивление космоцилину: [color=orange]Среднее[/color]
+diagnoser-disease-report-cureresist-high = Сопротивление космоцилину: [color=red]Сильное[/color]
+diagnoser-no-cures = Для это болезни отсутвуют особые лекарства.
+diagnoser-cure-has = Против заболевания имеються следующие лекарства:
+diagnoser-cure-bedrest = Лежать в кровати {$time} секунд, или {$sleep} на сон.
+diagnoser-cure-reagent = Следует принимать {$reagent} в объёме {$units}u.
+diagnoser-cure-wait = Пройдёт само через {$time} секунд.
+diagnoser-cure-temp = Достигните температуру тела ниже {$max}°K или выше чем {$min}°K.
+diagnoser-cure-temp-min = Держите температуру выше {$min}°K.
+diagnoser-cure-temp-max = Держите температуру ниже {$max}°K.
