+++ Resources/Locale/ru-RU/ganimed/disease/disease-proto.ftl
@@ -1,0 +1,24 @@
+disease-proto-furry-virus = Фуррфуиррус
+disease-meow = МЯУ!
+
+disease-proto-ultragigacancer = УЛЬТРАГИГАрак
+disease-proto-spectral-tiredness = фантомная усталость
+disease-proto-lung-cancer = Рак легких стадии IIIA
+
+disease-proto-space-cold = космическая изморозь
+disease-proto-vent-cough = вентиляционный кашель
+disease-proto-space-flu = космогрип
+disease-proto-bird-flew = птичий грип
+disease-proto-robovirus = Робовирус Ван Воусдаля
+disease-proto-amiv = AMIV
+disease-proto-amirmir = Мементик Амирмир
+disease-proto-bleeders = Укус Кровопотерь
+disease-proto-plague = Чума
+disease-proto-owonavirus = OwOнавирус
+disease-proto-slimevirus = Слаймовирус
+disease-proto-tongue-twister = Языковой Твистер
+disease-proto-dwarfpunch = Дварфийский Удар
+disease-proto-uraniumplague = Урановая Чума
+
+disease-proto-xeno = Ксеновирус
+disease-proto-xenodrone = Ксеновирус
