+++ Resources/Locale/ru-RU/ganimed/disease/disease.ftl
@@ -1,0 +1,12 @@
+﻿disease-cured = Вы чувствуете себя лучше.
+disease-sick-generic = Вы чувствуете себя больным.
+disease-eaten-inside = Вы чувствуете, как вас едят изнутри.
+disease-banana-compulsion = Вы очень реально хотите перекусить бананами!
+disease-beat-chest-compulsion = {CAPITALIZE(THE($person))} бьёт {POSS-ADJ($person)} торс.
+disease-vomit = {CAPITALIZE(THE($person))} рвёт.
+disease-think = Вы чувствуете, что не можете ясно мыслить.
+disease-polymorph = Вы чувствуете, как ваше тело скручивается и меняет форму!
+disease-sick-stomach = Вы чувствуете, что вас стошнит.
+
+disease-seizures-compulsion = Всё ваше тело непроизвольно дёргается.
+disease-blood-compulsion = Ваша кровь сгущается...
