+++ Resources/Locale/ru-RU/ganimed/disease/miasma.ftl
@@ -1,0 +1,4 @@
+﻿miasma-smell = Что-то дурно пахнет!
+miasma-rotting = [color=orange]Это гниёт![/color]
+miasma-bloated = [color=orangered]Оно раздулось![/color]
+miasma-extremely-bloated = [color=red]Оно сильно раздулось![/color]
