+++ Resources/Locale/ru-RU/ganimed/disease/scanner.ftl
@@ -1,0 +1,4 @@
+﻿disease-scanner-diseased = ЗАРАЖЁН
+disease-scanner-not-diseased = Нет заболеваний
+disease-scanner-gave-other = Ты дал {THE($target)} {CAPITALIZE($disease)}!
+disease-scanner-gave-self = Ты дал самому себе {CAPITALIZE($disease)}! поздравляем!
