+++ Resources/Locale/ru-RU/ganimed/disease/swab.ftl
@@ -1,0 +1,8 @@
+﻿swab-already-used = Вы уже использовали эту палочку.
+swab-swabbed = Вы берёте мазок рта у {THE($target)}.
+swab-mask-blocked = {CAPITALIZE(THE($target))} нужно снять {THE($mask)}.
+swab-used = Похоже, его использовали.
+swab-unused = Он чистый и готов к использованию.
+
+botany-swab-from = Вы бережно собираете пыльцу с растения.
+botany-swab-to = Вы тщательно высыпаете пыльцу на растение.
