+++ Resources/Locale/ru-RU/ganimed/disease/vaccine.ftl
@@ -1,0 +1,17 @@
+﻿vaxx-already-used = Вы уже использовали эту вакцину.
+vaxx-used = Оно потрачено.
+vaxx-unused = Оно небыло потрачено.
+
+vaccine-name = Вакцина: {$disease}
+vaccine-desc = Введя это, вы не сможете заразиться {$disease}, если у вас ее еще нет.
+
+vaccine-inject-start-agent = Вы начали вкалывать в {THE($target)} {THE($vaccine)}.
+vaccine-inject-start-patient = {CAPITALIZE(THE($user))} начал вкалывать в вас {INDEFINITE($vaccine)} {$vaccine}.
+
+vaccine-machine-menu-title = Вакцинатор
+vaccine-machine-menu-known-diseases-label = Известные заболевания
+vaccine-machine-menu-create-vaccine-button = Произвести вакцину
+vaccine-machine-menu-biomass-cost = Стоимость биомассой: {$value}
+vaccine-machine-menu-biomass-current = Содержится: {$value}
+vaccine-machine-menu-none-selected = Выберите болезнь!
+vaccine-machine-cost-upgrade = Цена биомассы
