+++ Resources/Prototypes/Entities/Clothing/Hands/gloves.yml
@@ -143,6 +143,8 @@
     sprite: Clothing/Hands/Gloves/latex.rsi
   - type: Clothing
     sprite: Clothing/Hands/Gloves/latex.rsi
+  - type: DiseaseProtection
+    protection: 0.1
   - type: Fiber
     fiberMaterial: fibers-latex
   - type: FingerprintMask
@@ -159,6 +161,8 @@
     sprite: Clothing/Hands/Gloves/alter_nitrile.rsi
   - type: Fiber
     fiberMaterial: fibers-nitrile
+  - type: DiseaseProtection
+    protection: 0.15
   - type: FingerprintMask
 
 - type: entity
