+++ Resources/Prototypes/Entities/Clothing/OuterClothing/bio.yml
@@ -9,6 +9,8 @@
     sprite: Clothing/OuterClothing/Bio/general.rsi
   - type: Clothing
     sprite: Clothing/OuterClothing/Bio/general.rsi
+  - type: DiseaseProtection
+    protection: 0.5
   - type: Armor
     modifiers:
       coefficients:
