+++ Resources/Prototypes/Entities/Mobs/Species/base.yml
@@ -237,6 +237,7 @@
   id: BaseMobSpeciesOrganic
   abstract: true
   components:
+  - type: DiseaseCarrier
   - type: Flashable
   - type: Barotrauma
     damage:
