+++ Resources/Prototypes/Entities/Mobs/Species/slime.yml
@@ -110,7 +110,12 @@
       types:
         Asphyxiation: -1.0
     maxSaturation: 15
-
+  - type: DiseaseCarrier
+    naturalImmunities:
+    - Ultragigacancer
+    - BleedersBite
+    - SlimeVirus
+    
 - type: entity
   parent: MobHumanDummy
   id: MobSlimePersonDummy
