+++ Resources/Prototypes/Entities/Structures/Machines/Medical/disease_diagnoser.yml
@@ -14,6 +14,12 @@
     - state: unlit
       shader: unshaded
       map: ["enum.DiseaseMachineVisualLayers.IsOn"]
+  - type: DiseaseDiagnoser
+  - type: DiseaseMachine
+    machineOutput: DiagnosisReportPaper
+  - type: DiseaseMachineVisuals
+    idleState: icon
+    runningState: running
   - type: Appearance
   - type: Machine
     board: DiagnoserMachineCircuitboard
