+++ Resources/Prototypes/Entities/Structures/Machines/Medical/vaccinator.yml
@@ -14,6 +14,8 @@
     - state: unlit
       shader: unshaded
       map: ["enum.DiseaseMachineVisualLayers.IsOn"]
+  - type: DiseaseMachine
+    machineOutput: Vaccine
   - type: Appearance
   - type: DiseaseMachineVisuals
     idleState: icon
@@ -24,3 +26,17 @@
     containers:
       machine_board: !type:Container
       machine_parts: !type:Container
+  - type: ResearchClient
+  - type: DiseaseVaccineCreator
+  - type: ActivatableUI
+    key: enum.VaccineMachineUiKey.Key
+  - type: ActivatableUIRequiresPower
+  - type: UserInterface
+    interfaces:
+      enum.VaccineMachineUiKey.Key:
+        type: VaccineMachineBoundUserInterface
+      enum.ResearchClientUiKey.Key:
+        type: ResearchClientBoundUserInterface
+  - type: MaterialStorage
+    materialWhiteList:
+    - Biomass
