+++ Resources/Prototypes/Entities/Structures/Machines/research.yml
@@ -12,6 +12,7 @@
       - state: server_o
         map: ["enum.WiresVisualLayers.MaintenancePanel"]
   - type: ResearchServer
+  - type: DiseaseServer
   - type: TechnologyDatabase
     supportedDisciplines:
     - Industrial
