+++ Resources/Prototypes/Ganimed/Deseases/noninfectious.yml
@@ -1,0 +1,64 @@
+﻿- type: disease
+  id: Ultragigacancer
+  name: disease-proto-ultragigacancer
+  infectious: false
+  cureResist: 0.15
+  effects:
+  - !type:DiseaseHealthChange
+    probability: 0.3
+    damage:
+      types:
+        Cellular: 1
+  - !type:DiseaseVomit
+    probability: 0.01
+  - !type:DiseasePopUp
+    probability: 0.03
+  cures:
+  - !type:DiseaseReagentCure
+    reagent:
+      ReagentId: Phalanximine
+    min: 15
+
+- type: disease
+  id: SpectralTiredness
+  name: disease-proto-spectral-tiredness
+  infectious: false
+  effects:
+  - !type:DiseaseGenericStatusEffect
+    probability: 0.03
+    key: ForcedSleep
+    component: ForcedSleeping
+    time: 3
+    type: Add
+  - !type:DiseaseSnough
+    probability: 0.025
+    emote: Yawn
+  - !type:DiseaseHealthChange
+    probability: 0.02
+    damage:
+      types:
+        Poison: 4
+  cures:
+  - !type:DiseaseJustWaitCure
+    maxLength: 240
+  - !type:DiseaseBedrestCure
+    maxLength: 60
+
+- type: disease
+  id: StageIIIALungCancer
+  name: disease-proto-lung-cancer
+  infectious: false
+  cureResist: 1.0
+  effects:
+  - !type:DiseaseHealthChange
+    probability: 0.3
+    damage:
+      types:
+        Cellular: 1
+  - !type:DiseaseVomit
+    probability: 0.01
+  - !type:DiseaseSnough
+    probability: 0.10
+    emote: Cough
+  - !type:DiseasePopUp
+    probability: 0.03
